requirements.txt
tensorflow==2.20.0
keras==3.12.1
flask==3.1.3
opencv-python==4.10.0.84
numpy==2.2.6
matplotlib==3.10.8
pillow==12.1.1
psycopg2-binary==2.9.11
werkzeug==3.1.3
